fix: props 전달값 오류 해결 (#36)
